Add github webhook
This is a modified version of the webhook used for a couple of our
auto-deploys:

 - https://github.com/ocf/puppet/blob/281f0c13/modules/ocf_puppet/files/webhook/github.cgi
 - https://github.com/ocf/puppet/blob/4a926702/modules/ocf_wiki/files/webhook/github.cgi

This should allow us to re-use the code without repeating it in each
webhook we need.

I was a little stuck on where to put this (considered puppet repo
instead of utils), but decided utils was better. While it's not a
perfect fit (it's not a utility someone would ever execute), I don't
really think there's a better place to put it. I don't like using the
puppet repo for managing scripts, so utils seems like a reasonable
location. It wouldn't necessarily hurt to put this in its own repo, but
I don't know what else we would ever put there...