Reorganize scripts, remove system-only scripts
To be consistent, the new policy is that scripts in utils should be
intended primarily for human use, and scripts intended for system use
(cronjobs, puppet exec, etc.) should go in puppet. This still isn't a
perfect way to determine what goes where, but it's better than before
(having no guidelines).

This also moves some scripts into appropriate directories, and removes
some file extensions, also to be consistent with the new README.