Rename {paper => oldpaper}, {paper2 => paper}
Reading from the enforcer DB will be the only way to get an accurate
page count now that some people will be using dev-pollution which only
logs to enforcer.

Since pollution logs to both enforcer and PyKota, it's okay to use
enforcer as the authoritative source in all cases.

"oldpaper" is still around for now in case staff want to use it for
debugging, but it will be removed in the near future.