echo stty after ocf-tv connect
If ssh is killed during a password prompt, terminal echo is
disabled and needs to be re-enabled. This should not happen
with cleanup after SIGTERM, but just in case, per #32.

Alternatively, we can disable ssh password prompts entirely
with `-o BatchMode=yes` and use Kerberos/GSSAPI.
Personally I prefer that approach to providing my password
to an ocf-tv script, but this is a simple change for now.