minutes: distinguish directors and guests in BoD attendance lists
This commit makes the minutes script split the "Attendance"
block in BoD minutes. The note-taker still writes a single
Attendance block, like before; however, after they exit the
text editor, the Attendance block is replaced with two blocks,
an attendance block for directors and an attendance block for
guests. Guests are further marked with an asterisk if they
choose to join BoD at that meeting.

Major changes have been made to the BoD module. Most notably,
the membership file is eliminated. Membership is now dynamically
determined from the attendance lists in the minutes.

One drawback is that you cannot run `minutes` to take BoD minutes
twice in the same day without re-combining the attendance lists,
or else it crashes. This can be remedied in the future.