Made changes to easywp script.
Added a description to the top of the file. Added pipefail with euxo flags to the
script to make it less dangerous. Added globbing support for hidden files. If a user
has files in their webroot, it makes a backup in their home directory. Added requesed
error messages and informative echo statements.