Find TV output by card name
Since the sink name depends on the profile the card is in, search for
the TV's output sink by card name. The Python function to return the
sink is currently unused, but kept in case the sink name is ever needed
outside the context of a command running over ssh. Also fix a typo in a
help string.