how: use bash
Dash does not understand `command -v` with multiple parameters, which is
causing the `how` command to fail with the output

```
/opt/share/utils/bin/how: 63: /opt/share/utils/bin/how: -en: not found
```

because it tries to execute `$(command -v greadlink readlink) -en ...`
but `command` isn't outputting anything.

Furthermore, converting this script to a bash script makes it more consistent
with how other OCF shell scripts work.